requirements.txt
pyyaml>=6.0
pillow>=9.0.0
requests>=2.28.0
pytest>=7.0.0
pytest-mock>=3.10.0 